+++ config.ts
@@ -8,6 +8,7 @@ const config: Manifest = {
   homepage: pkg.homepage,
   license: pkg.license,
   author: pkg.author,
+  target_engine: '1.1.0',
   categories: [],
   tags: [],
   download_url_template: 'https://github.com/any-listen/any-listen-extension-online-metadata/releases/download/v{version}',
@@ -16,13 +17,8 @@ const config: Manifest = {
   contributes: {
     resource: [
       {
-        id: 'kw',
-        name: '{kwName}',
-        resource: ['musicSearch', 'musicPic', 'musicLyric'],
-      },
-      {
-        id: 'tx',
-        name: '{txName}',
+        id: 'kg',
+        name: '{kgName}',
         resource: ['musicSearch', 'musicPic', 'musicLyric'],
       },
       {
@@ -31,8 +27,13 @@ const config: Manifest = {
         resource: ['musicSearch', 'musicPic', 'musicLyric'],
       },
       {
-        id: 'kg',
-        name: '{kgName}',
+        id: 'kw',
+        name: '{kwName}',
+        resource: ['musicSearch', 'musicPic', 'musicLyric'],
+      },
+      {
+        id: 'tx',
+        name: '{txName}',
         resource: ['musicSearch', 'musicPic', 'musicLyric'],
       },
     ],